Rewrote a lot of the network parts and a bit of code
Created a prefab that both pipe and tank extend so they share the same
network code. This should decrease issues with errors that arise between
the two. As well i reworked a few things in the network fluid tile class
to prevent issues.

+++ src/dark/fluid/client/model/ModelLiquidTank.java
@@ -359,7 +359,7 @@ public class ModelLiquidTank extends ModelBase
 
     public void renderMeter(TileEntity tee, float f5)
     {
-        int[] conenctedTiles = new int[6];
+        byte[] conenctedTiles = new byte[6];
         if (tee instanceof TileEntityTank)
         {
             conenctedTiles = ((TileEntityTank) tee).renderConnection;
